refactor(common-model-converters): improve re-usability of ModelConverters
extracted SimpleClassModelConverter
